Moved tile stuff into package
